Moving files - Added a models folder in src - Moved migration to a folder at the root of the project
