+++ knexfile.js
@@ -1,4 +1,5 @@
 // Update with your config settings.
+// Keep it as JS because it's useless to make it a TS file and it has to be at the root of the project
 import * as Knex from "knex";
 const config = require("./config.json");
 
@@ -14,11 +15,11 @@ export default {
             user : config.testdatabaseuser,
         },
         migrations: {
-            directory: __dirname + "src/db/migrations",
+            directory: __dirname + "/migrations",
             tableName: "knex_migrations",
         },
         seeds: {
-            directory: __dirname + "src/db/seeds/development",
+            directory: __dirname + "/seeds/development",
         },
 
     },
@@ -33,7 +34,7 @@ export default {
             user:     config.productiondatabaseuser,
         },
         migrations: {
-            directory: __dirname + "src/db/migrations",
+            directory: __dirname + "/migrations",
             tableName: "knex_migrations",
         },
         pool: {
@@ -41,7 +42,7 @@ export default {
             min: 2,
         },
         seeds: {
-            directory: __dirname + "src/db/seeds/production",
+            directory: __dirname + "/seeds/production",
         },
     },
